Fix most of #153 issues on SFX-8 branch (Implicit conversions SFX to/from JFX should allow `null` arguments)
There is an inconsistency, comapred toth rest of the code, in implementation of SceneAntialiasing JFX to SFX conversion.
When argument is null it returns `Disabled`. While it may be nice in this case it si inconsistant with the rest of the API.
Right now test "A SceneAntialiasing should allow `null` as an argument to implicit conversion from JFX to SFX" fails.