Correct fix for Issue #169 - do not assume that "update" event contains full information about the changes.
The documentation for the "update" event does not state that there is information about the content that changes, only its location (from, to).
It cannot be assumed that `getAddedSubList` contains the update. The behaviour changed between JavaFX 2 and 8.
Provide fix that is in line with JavaFX 2 documantation but also works in JavaFX 8, so there are no uncessary diferences between versions.