Fixes #190: Convenience methods for clipboard assignment don't work
Because convenience methods such as `content_=` require a method without
the `_=` as part of the name, I've modified content to return all the
data on clipboard as `ClipboardContent`.

The individual `getContent` method from JavaFX has been renamed to get,
and optionally returns an object for the given data format, instead of
returning null, in order to be null-safe.