Fix for unable to pass lambda into KeyFrame event handler
The KeyFrame code strongly relies on being able to receive an
EventHandler rather than an actual lambda. This means that an implicit
conversion from a lambda into an EventHandler is absolutely necessary,
and the lack of such a conversion would result in a lot of changes
across the codebase.

As a fix, I have restored the original implicit conversion. The caveat
now is that lambda assignment must type annotate the lambda parameters,
which is what this branch set out to do in the first place. For now,
lambda assignment is introduced in a backward compatible manner, with
the only minor improvement being that no implicit conversion is required
at the user level to assign lambdas.