Fix TextFieldListCell.converter_=(c) method
Fixes #262. The `converter()` method of the `ConvertableCell` trait
was creating a fresh ScalaFX `ObjectProperty` out of the ***value***
of underlying `converterProperty()`. While it is OK when reading the
converter value, it didn't work for ***setting*** new converter through
this wrapper from ScalaFX (this can still be achieved through the
`delegate.setConverter(c)`).

This commit changes the `converter()` method to return a wrapper
for the underlying ***property*** instead of value.

This change technically can break some existing code, though,
since it changes the type of the `converter()` method.