Update Color.scala
SFX Color.grayRgb is used to wrap the grayRgb method in JavaFX. Unlike JFX's gray method, grayRgb is of type Int and has a range of 0 - 255, which corresponds to 0.0 - 1.0 of the Double type of gray. The current SFX delegates grayRgb directly to gray, and the Int type is converted and automatically converted to Double, which is wrong and cannot be checked by the checkRGB method.